Change of CBOR library. Cleaning up some build system cruft.
