+++ Makefile
@@ -50,7 +50,6 @@ LIBS = -L$(OUTPUT_PATH) -L$(WHERE_I_AM)/lib -lstdc++ -lm -lmanuvr
 # Wrap the include paths into the flags...
 CFLAGS += $(OPTIMIZATION) $(INCLUDES)
 CFLAGS += -fsingle-precision-constant -Wdouble-promotion
-CFLAGS += -fno-rtti -fno-exceptions
 
 
 ###########################################################################
@@ -104,7 +103,7 @@ MANUVR_OPTIONS += -DMANUVR_GPS_PIPE
 # Wire and session protocols...
 #MANUVR_OPTIONS += -DMANUVR_SUPPORT_OSC
 MANUVR_OPTIONS += -DMANUVR_OVER_THE_WIRE
-MANUVR_OPTIONS += -DMANUVR_SUPPORT_MQTT
+#MANUVR_OPTIONS += -DMANUVR_SUPPORT_MQTT
 MANUVR_OPTIONS += -DMANUVR_SUPPORT_COAP
 MANUVR_OPTIONS += -DMANUVR_CBOR
 MANUVR_OPTIONS += -D__MANUVR_CONSOLE_SUPPORT
@@ -136,7 +135,8 @@ LIBS += $(OUTPUT_PATH)/libmbedcrypto.a
 # Merge our choices and export them to the downstream Makefiles...
 CFLAGS += $(MANUVR_OPTIONS)
 export CFLAGS
-export CPP_FLAGS = $(CFLAGS)
+export CPP_FLAGS    = $(CFLAGS) -fno-rtti -fno-exceptions
+
 
 .PHONY: all
 
@@ -167,13 +167,13 @@ debug: clean libs
 tests: libs
 	export __MANUVR_LINUX
 	$(MAKE) -C ManuvrOS/
-	$(CXX) -static -g -o dstest tests/TestDataStructures.cpp $(CFLAGS) -std=$(CPP_STANDARD) $(LIBS) -D_GNU_SOURCE -O2
-	$(CXX) -static -g -o bptest tests/BufferPipeTest.cpp $(CFLAGS) -std=$(CPP_STANDARD) $(LIBS) -D_GNU_SOURCE -O2
+	$(CXX) -static -g -o dstest tests/TestDataStructures.cpp $(CPP_FLAGS) -std=$(CPP_STANDARD) $(LIBS) -D_GNU_SOURCE -O2
+	$(CXX) -static -g -o bptest tests/BufferPipeTest.cpp $(CPP_FLAGS) -std=$(CPP_STANDARD) $(LIBS) -D_GNU_SOURCE -O2
 
 examples: libs
 	export __MANUVR_LINUX
 	$(MAKE) -C ManuvrOS/
-	$(CXX) -static -g -o gpstest examples/tcp-gps.cpp $(CFLAGS) -std=$(CPP_STANDARD) $(LIBS) -D_GNU_SOURCE -O0
+	$(CXX) -static -g -o gpstest examples/tcp-gps.cpp $(CPP_FLAGS) -std=$(CPP_STANDARD) $(LIBS) -D_GNU_SOURCE -O0
 
 builddir:
 	mkdir -p $(OUTPUT_PATH)
