+++ lib/Makefile
@@ -1,5 +1,6 @@
 # Build all libraries
-C_STANDARD = c99
+C_STANDARD   = c99
+CPP_STANDARD = gnu++11
 
 # Oh man.... this will be very messy.
 
@@ -16,23 +17,25 @@ SOURCES_MQTT_C += paho.mqtt.embedded-c/MQTTPacket/src/MQTTSubscribeServer.c
 SOURCES_MQTT_C += paho.mqtt.embedded-c/MQTTPacket/src/MQTTUnsubscribeClient.c
 SOURCES_MQTT_C += paho.mqtt.embedded-c/MQTTPacket/src/MQTTUnsubscribeServer.c
 
-SOURCES_CBOR_C  = tinycbor/src/cborerrorstrings.c
-SOURCES_CBOR_C += tinycbor/src/cborencoder.c
-SOURCES_CBOR_C += tinycbor/src/cborencoder_close_container_checked.c
-SOURCES_CBOR_C += tinycbor/src/cborparser.c
-SOURCES_CBOR_C += tinycbor/src/cborparser_dup_string.c
-SOURCES_CBOR_C += tinycbor/src/cborpretty.c
-SOURCES_CBOR_C += tinycbor/src/cbortojson.c
+SOURCES_CBOR  = cbor-cpp/src/encoder.cpp
+SOURCES_CBOR += cbor-cpp/src/decoder.cpp
+SOURCES_CBOR += cbor-cpp/src/input.cpp
+SOURCES_CBOR += cbor-cpp/src/listener_debug.cpp
+SOURCES_CBOR += cbor-cpp/src/output_dynamic.cpp
+SOURCES_CBOR += cbor-cpp/src/output_static.cpp
 
-SOURCES_C  = $(SOURCES_MQTT_C) $(SOURCES_CBOR_C)
-SOURCES_C += $(SOURCES_UECC_C)
+SOURCES_C   = $(SOURCES_MQTT_C)
+SOURCES_C  += $(SOURCES_UECC_C)
+
+SOURCES_CPP = $(SOURCES_CBOR)
 
 FLAGS_LIBMQTT  =
 
-LOCAL_C_FLAGS = -nostdlib $(FLAGS_LIBMQTT)
+LOCAL_C_FLAGS   = -nostdlib $(FLAGS_LIBMQTT)
+LOCAL_CXX_FLAGS = -nostdlib $(CPP_FLAGS)
 
 
-OBJS  = $(SOURCES_C:.c=.o)
+OBJS  = $(SOURCES_C:.c=.o) $(SOURCES_CPP:.cpp=.o)
 LIBNAME = extraprotocols.a
 
 
@@ -44,8 +47,10 @@ all: $(LIBNAME)
 #	$(MAKE) -C telehash-c
 
 %.o : %.c
-	$(CC) $(INCLUDES) $(CFLAGS) $(LOCAL_C_FLAGS) -std=$(C_STANDARD) -nostdlib -c $< -o $@
+	$(CC) $(INCLUDES) $(LOCAL_C_FLAGS) -std=$(C_STANDARD) -c $< -o $@
 
+%.o : %.cpp
+	$(CXX) $(INCLUDES) $(LOCAL_CXX_FLAGS) -std=$(CPP_STANDARD) -c $< -o $@
 
 $(LIBNAME): $(OBJS)
 	$(AR) -rvs $(OUTPUT_PATH)/$@ $(OBJS)
